# public/openapi.yaml
openapi: 3.1.0
info:
  title: Conduit API Gateway
  version: 1.0.0
  description: |
    Conduit is a transparent, edge-first API gateway and proxy for Claude AI.

    ## Features
    - API Key Management with rotation and expiration
    - Multi-provider support with automatic failover
    - Rate limiting and quota enforcement
    - IP whitelisting and HMAC request signing
    - Response caching for improved performance
    - Multi-tenancy with organizations and projects
    - Comprehensive analytics and cost projections
    - Audit logging and 2FA authentication

  contact:
    name: Conduit Support
    url: https://conduit.dev
  license:
    name: MIT
servers:
  - url: https://api.conduit.dev
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/admin/keys/{id}/rotate:
    post:
      tags: [API Keys]
      summary: Rotate API key
      description: Generates a new key while preserving all settings
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Key rotated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  newKey:
                    type: string
                  rotatedAt:
                    type: string
                    format: date-time
...
  /api/admin/analytics/projections:
    get:
      tags: [Analytics]
      summary: Get cost projections
      security:
        - BearerAuth: []
      parameters:
        - name: apiKeyId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cost projection data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostProjection'
  /api/admin/analytics/errors:
    get:
      tags: [Analytics]
      summary: Get error rate statistics
      security:
        - BearerAuth: []
      parameters:
        - name: apiKeyId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Error rate statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalRequests:
                    type: integer
                  successfulRequests:
                    type: integer
                  failedRequests:
                    type: integer
                  errorRate:
                    type: number
                    format: float
                  errorsByCode:
                    type: array
                    items:
                      type: object
                      properties:
                        statusCode:
                          type: integer
                        count:
                          type: integer
  /api/admin/analytics/latency:
    get:
      tags: [Analytics]
      summary: Get latency statistics
      security:
        - BearerAuth: []
      parameters:
        - name: apiKeyId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Latency statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  averageLatencyMs:
                    type: integer
                  medianLatencyMs:
                    type: integer
                  p95LatencyMs:
                    type: integer
                  p99LatencyMs:
                    type: integer
                  minLatencyMs:
                    type: integer
                  maxLatencyMs:
                    type: integer
  /api/admin/audit-logs:
    get:
      tags: [Audit Logs]
      summary: Get audit logs
      security:
        - BearerAuth: []
      parameters:
        - name: resourceType
          in: query
          schema:
            type: string
        - name: resourceId
          in: query
          schema:
            type: string
            format: uuid
        - name: adminId
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Conduit API key (sk-cond_...)
    HmacAuth:
      type: apiKey
      in: header
      name: x-signature
      description: HMAC signature for request signing
    BearerAuth:
      type: http
      scheme: bearer
      description: Admin session token
  schemas:
...
    CostProjection:
      type: object
      properties:
        currentSpend:
          type: integer
          description: Current month spend in cents
        projectedMonthlySpend:
          type: integer
          description: Projected end-of-month spend in cents
        projectedDailyAverage:
          type: integer
          description: Daily average spend in cents
        trend:
          type: string
          enum: [increasing, decreasing, stable]
        percentageChange:
          type: number
          description: Percentage change vs previous month
        daysInPeriod:
          type: integer
        estimatedEndOfMonthSpend:
          type: integer
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        adminId:
          type: string
          format: uuid
        adminEmail:
          type: string
        resourceType:
          type: string
        resourceId:
          type: string
          format: uuid
        action:
          type: string
        changes:
          type: object
        ipAddress:
          type: string
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer